Have tests use tasty.
This is implemented by having distributed-process-tests implemented
on top of tasty rather than Cabal.

The API changes are in withTmpDirectory which works over IO computations
rather than tests, and defaultMainWith which now requires tasty
arguments to be separated from tests with a '--' separator. Ideally we
should use tasty options to define custom command line arguments, this
is still pending.