Add an explicit event route tracking.
This commit introduces an additional field to the HAEvent datatype
it allow tracking station to track the path that HAEvent follows.
Original NodeAgent set path to [], and each EQ that is on the path
add it's address to the route.

The final EQ adds it's address to the path when send message to the
RC, but not to the replicated state, this allow us to change address
uppon a recovery.

Current implementation it not full as currently multiple EQ are not
supported and small changes to the EQ procedure may be implemented.