Capture test output to files.
Firstly, we include the tasty-files submodule and change existing test runners to use it. tasty-files rewrites the test tree to add a pseudo-resource block around each test, which redirects stdout and stderr to files. The existing console test runner is then used to execute each test.

tasty-files can capture stdout and stderr, but capturing stdout seems to nondeterministically capture small parts of the test runner output as well as the test output; at the moment, I have switched this off, which means that small amounts of stdout still get printed to the screen. All of stderr is captured correctly, however.

The second change alters GHC event tracing to output to files rather than stderr. When event tracing is enabled, tasty-files delimits each test with markers in the event log. Ideally, we would be able to output the event log messages to stderr and capture them as normal, but the event log lacks an accessible flushing mechanism, meaning we cannot make sure that messages get output in the correct test file (or in a file at all).