Adjust HA.NodeAgent.handleEQResponse to staggering.
The check that was used to determine if the preferred node has been
tested was assuming the serialCall behavior.

With staggering, we assume the preferred node has been contacted if
the responding replica was contacted in the second phase of staggering.
This translates to asking if the responding replica is at the head of
the list of replicas or not.