replicated-log: Introduce a current unconfirmed decree.
The old current decree is now called the current unconfirmed decree,
which is the decree identifier of the next proposal to confirm. All
previous decrees are known to have passed consensus. The current
decree is now the next proposal to submit.

This change allows to submit proposals one by one to the proposer
process, avoiding submitting multiple proposals with the same decree
from the same replica.