Use our own vendor/clock submodule.
Upstream had numerous, numerous bugs. For example, GHC Int's were used
to model C 64-bit signed integers, even though GHC does not guarantee
the full range that C does. Another problem was that the size of
TimeSpec was architecture dependent. Another was that it assumed time_t
= long, when on modern 32-bit systems time_t is actually long long.

Upstream also does not yet have TimeSpec arithmetic. We don't want to be
implementing that ourselves, and get it wrong (in fact we have!). The
vendored submodule contains fixes for all of the above, cleanups and
TimeSpec arithmetic.