Proper handling of build -> dep dependency in the Makefile
By enabling the DELETE_ON_ERROR feature of make, we can use the
cabal.sandbox.config file as a stamp file on the successfulness
of building the sandbox.  This means that after a successful
`make dep' invocation, the second `make dep' is a quick no-op.

Therefore we can make build depend on the dep target, properly
enforcing an invariant.