Use the `Dict` type from the constraints package in the `Resource`/`Relation` classes.
Summary:
As part of this change, the `Dict` is no longer existentially
wrapped in `resourceDict` and `relationDict`. This means extra static
checking. Previously, it was possible to pass to define a `Resource`
instance using the wrong dictionary. Also, not existentially wrapping
means resource definition became a whole lot simpler: no need for any
proxy arguments anymore.

Also, we were doing some funny business previously in the `Relation` type class. In order to use a single `Some` datatype for existential wrapping, we shoe horned three type params into a single tuple of three type params. This is dangerously confusing, so this patch cleans that up.

Using the Dict datatype from `constraints`  package is prep work for transitioning to the `(:-)` type that we will ultimately need, as explained in a recent email to the list.

Test Plan: `make build`

Reviewers: mboes

Reviewed By: mboes

Subscribers: facundominguez

Differential Revision: http://178.62.3.60:8081/D7