Initial implementation of bindings to libgenders.
Summary:
It turns out genders is not a bad format for configuring
nodes in a cluster. Chief advantage: it makes it easy to apply common
sets of attributes to large sets of nodes, rather than have to repeat
those attributes for each node (and have them potentially go out of
sync). Might as well reuse it, especially given that it's already an
established format in this space. We could of course preprocess
a genders file out-of-band using `nodeattr` and pipe parameters via
stdout, but that's brittle, would probably amount ultimately to
implement some ad hoc parameter parser redoing much of what libgenders
does, and it's not like the libgenders library is that big or
complicated. Bindings should be much more robust, and gives us some
extra type safety.

So I dusted up some old code (from this summer maybe?) and created
a test suite for it. Discovered `hspec` for this, and it's great. Have
a look at the test suite.

This is a fairly complete set of bindings. All libgenders functions are
made available through `Database.Genders.Internal`. A mostly pure
interface to genders is provided in `Database.Genders`.

Using small genders file should be useful for d-commands, in
particular whenever different nodes have different roles.

Test Plan: `cabal test --show-details=always` in `genders/`.

Reviewers: nclarke

Reviewed By: nclarke

Subscribers: facundominguez

Differential Revision: http://178.62.3.60:8081/D14