Remove 'Address' and 'Network' abstractions.
Summary:
These were broken abstractions. For one 'Address' was leaky,
because it exposed implementation details which e.g. Lookup.hs was
fully aware of. Second, it was adding yet another "address" type,
layered on top of all the existing ones, when really whatever is in
network-transport should suffice. In the case of TCP, 'Network' added
extra structure to 'Transport's, but this extra structure was only
useful for the 'networkBreakConnection' call, which itself was only
useful in tests. We shouldn't be adding layers on top of 'Transport':
those unit tests that need connection breaking should allocate the
'Transport' themselves and ask for the TCP internals to be exposed.

In the end, the abstraction was adding more code than it saved: removing
the abstraction saves us 70 lines net. In fact even more lines should
disappear in the future, when we remove the node agent lookup mechanism,
and push using `NodeId`s as early as possible in the control flow as
possible.

RPC changes on a best effort basis, seeing as we can't test this
currently. But I believe I haven't broken anything significant.

Test Plan: `make build`.

Reviewers: facundominguez, nclarke

Reviewed By: nclarke

Differential Revision: http://178.62.3.60:8081/D12