replicated-log: Pack user tunable "config" arguments into a Config structure.
The signature to 'replica' was getting a bit long (... 11 arguments!),
so I packed up most arguments into a Config data structure, to which we
will be able to add fields to our heart's content.

'new' now has 3 useful arguments: the config, a state machine
definition, and the set of nodes to spawn on (i.e. the initial
membership). That's a pretty straightforward signature. The state
machine definition (i.e. Log) is still not part of Config, but that's
because it's morally distinct: the Log structure states *what* the
replicated state machine does, the Config structure states *how* (e.g.
where to store data, how long leases are for, etc).

This change makes using replicated-log arguably quite a bit less verbose
and therefore much more straightforward. See tests for an illustration
of that.