Rewrite the RG representation of services.
A service is now represented by its ProcessId. This is another step
towards removing the NodeAgent, since we have a central repository
to use in place of the node-local registry.

Note that this patch does not attempt to simplify the message types
(and remove ProcessEncode where possible); we leave this to a subsequent
patch.