replicated-log: Introduce ReplicaState.
Pack the myriad arguments to the `go` auxiliary function in `replica`
into a record. This is more verbose in some ways, but it should reduce
merge conflicts and contribute to retaining sanity. At least with this,
we save on having to repeat all invariant arguments at each recursive
call - now we only have to write explicitly those that vary at that
particular recursive call site.

Note that the variable names stay the same as before. This is a half
measure that allows to not rewrite all variable names everywhere.