Remove the nodeagent from the startup process.
This is one step in the grand and slow process of removing the node
agent entirely, and corresponds to point 1 identified in
https://app.asana.com/0/12314345447678/16249819340711

This should introduce no breaking changes. Tests still pass. It should
be possible, in light of this commit, to run NA-less tracking stations,
provided they themselves do not need to host other services.