replicated-log: Workaround race in update-handle test.
After adding replicas to a group, the state takes a while to replicate.
If the old replicas are killed too soon, the new replicas will be unable
to do any progress.

The current workaround introduces a delay in the test so new replicas
have more time to replicate the state. This is a temporal solution that
needs a better fix.