replicated-log: Implement snapshots.
Summary: Persisting the membership together with the snapshots would simplify maintaining the invariant that the latest membership can be reconstructed from the latest snapshot and the log. But then the interface for logInitialization/dump/restore would become more noisy. For now the membership is stored together with the log.

Test Plan: make && CABAL_FLAGS="-fdistributed-tests --reinstall" make PACKAGES=mero-halon

Reviewers: mboes

Reviewed By: mboes

Subscribers: benclifford, nclarke, facundominguez, mboes

Differential Revision: http://178.62.3.60:8081/D28