halon: Retries requests for the replicated state if they take long.
In essence every call of the Replicator is interrupted after a while
using either one of the functions 'timeout' or 'retry', which are
introduced in the module "HA.System.Timeout".

This opens the possibility for abandoned requests to still happen
out of order and arbitrarily late, but this is something to address
in another commit.