halon: Fix race in RS where the RC could survive much beyond its lease.
This race was due to the ability to cancel a timer after it has
expired. The timer is programmed to kill the RC. The timer could be
canceled after the lease expires but before the programmed action
executed. In this case, the RC would be kept alive while a second
RC is elected somewhere else.

With this patch, timers are not canceled if the timer period expired.