Make this simpler.
Given the model we're using, there should never be a need to remove
edges, since the presence of an edge between a GC component and a non-GC
component would imply it's not ready to be garbage collected. Edges are
stored on nodes, so edges in the GC component will go anyway.