replicated-log: Initialize the legislatureId from snapshots.
Before this commit, starting the replica from a snapshot would make no
effort to figure out at which legislature the snapshot was taken. The
result being that an incorrect LegislatureId would be produced from
executing a trimmed log.

In this commit, LegislatureIds are saved together with the snapshots.
In addition, the log also stores the LegislatureId of the membership
recorded when the log is trimmed so it can be used for initialization.