Implement 'Interesting Event Message' channel.
This patch does a few things:
- Implements an 'interesting event message' type and the machinery for
  sending it to SSPL.
- Tidies the existing SSPL implementation to be more general.
- Uses channels for external communication.

This will require support in the RC to process the 'DeclareChannels'
message correctly and record this in the RC. This will need to wait for
Yorick's changes to the RC to do in a nice manner.

Existing functionality has been verified to still work.