d-p-scheduler: Fix race when puting messages in mailboxes.
It could happen that the scheduler asked a process to check its mailbox
before puting a message in it. This would cause the system to block
since the process would declare it cannot make progress and the
schedular would not have any other actions to take.

Typed-channels are not supported in the remote case.