Fix broken tests.
We now properly register the RC node as a 'node'. Previously it was
coming up as a sort of `pseudo-node`, with some of the aspects we
expect of a node but not all. This was confusing the tests for some
reason I haven't explicitly pinned down, but this fixes anyway.