Fix master monitor test.
The problem is that we would like to uniquely identify the master monitor with the
`MasterMonitor` link from the Clsuter node. But this doesn't get created for a
while, so we have another race condition with services starting. Hopefully
the new RC design should help solve this, but for now we use a fallback if
we can't find the `MasterMonitor` link, where we try to look up the service
directly on the current node.