Process IOS config files, add drive info into graph.
Actually, I'm not entirely convinced by this way of modelling drives.
The problem is that we have to separate them completely to deal with
the different data we get when adding them. I am thinking instead
of introducing a single 'Drive' concept which has a unique, Halon
generated ID. This would then be added to a number of other identifiers.
Drives could then be merged under certain rules when we understand that
they refer to the same piece of hardware. Still thinking about the
implications of this.