Rewrite 'StorageDevice' schema.
This removes the artificial distinction we previously had, and mostly
fixes the strange identification scheme for drives. Now drives are
uniquely identified by a Halon-internal ID, and may have other
identifiers attached. Depending on the context, one or more of these
other identifiers may be used to select a drive.