Change identification of `HAEvent`.
HAEvent was using a spurious identification scheme since the removal
of the node agent. This commit alters this and switches to using
UUIDs. We also bolster the EQ tests by explicitly checking event
IDs, which was missing in previous iterations.

This branch does not make further potential changes to `HAEvent`,
such as including a timestamp.

An additional function is now exposed by `Producer` in order to
support sending `HAEvent`s directly. This is used in testing.

Remove `EventId` entirely.