Introduce coverage scripts.
This commit introduces a coverage script that could be used
standalone or on a CI. This script creates .coverage directory,
compile libraries with relevant flags, gather output and generate
reports.
This script is a workaround to the different -hpc limitations
of cabal. For example cabal requite all dependencies to be build
with coverage support and fails to build all halon dependencies
with it.
All build results are stored as build artifacts and could be
analysed after run.

Changes in build system:
  * Unconditionaly generate hpc files: .mix, .tix for $PACKAGES
  * Install $PACKAGES one by one
  * $PACKAGES in mk/config.mk put into topological order