Initial stack support.
We no longer use submodules. Instead, we use stack's support for pulling
directly from git repos. This works well because the submodules we have
are largely vendor packages, which have an independent release cycle
from Halon. We can be hacking on them in a separate directory. We don't
need to be hacking on both vendor packages and Halon at the same time.

We take this opportunity to use release versions of network and
leveldb-haskell.