halon: Fix the RS timing.
With the previous algorithm it was possible for the RS to think it
had the lease when it really didn't. The lease could be obtained when
a timer was started, and then it could be renewed beyond the current
lease expiration.

Now it is ensured that lease renewal happens while keeping the lease
or otherwise the RC is killed.