Remove EQT from beign a service.
Because event queue is required process on the node
and have no configuration it makes no sense to make
it a service. Because making it a service introduces
a huge window for additional failures and race conditions
and forces higher level libraries implement bootstrap
procedure, that lead to a mixture of different layers.