Port mero-halon to new EQT API.
After removing EQT from beign a service RC rules are
simplified as we don't need to start EQT service.
Also we have added additional update of EQT on
NodeUp event, because node just started, and on respawn
of the Monitor service. Latter means that node just
connected to cluster and may be not up to date.