Run the halond from test-start-service-local inside temp dir
There are only certain filesystems that mero trace files can be mmaped
from. For example, it does not work if the test is ran from a vboxfs
host mount. Instead of creating test artifacts inside the build
directory somewhere, we use a temporary directory. This should mean
that the test works more often.

Note that this does not mean the test doesn't need manual
intervention. The user still has to make sure the mero kernel modules
are started first. There is room for improvement in error reporting
from the tests.