Simplify CEP framework implementation.
The general idea was to remove additional layers in CEP
implementation to make it's easier to maintain and reason
about. The general rule was that each SM should care either
about some local state that should not leak to layers above
or for some longrunning computation that should be delayed
eventually.
For this reason only 2 layers of State Machines left:
1. SM (State Machine), previously Stack, that run a computation
   and carry a buffer and local state;
2. Engine that run many statemachines and keep internal information
   about system state
Redundant layers: SM and StackDriver were removed.

List changes:
  * Global state was converted to an explicit output parameter
    and was removed from StackOut, SMOut and related types.
  * A notion of suspended and runnable rules was introduced.
    Suspended rule is a rule that can't process without more input.
  * If Engine has runnable rules caller do a non-blocking read
    and continue working with rules, otherwise it fallback to
    original implementation and do a blocking read.
  * SM was converted from State Machine to a PhaseM runner
    function and moved to Network.CEP.Phase module.
  * Switch and normal rules were generalized to one rule Stack
    was renamed to SM and moved to Network.CEP.SM module.