Add TH for auto-generating resources/relations.
This is messier than I would like, for two reasons:

- Trying to do everything in one go runs afoul of TH splicing
restrictions, and so two calls must be make; one to create the initial
dicts, and one to create the 'remotable' block and the
Resource/Relation instances.

- In order to support things like `Resource (Maybe String)`, we have
to use `[TypeQ]` rather than `Name` as the argument, and thus either
use a quasiquote or prepend `conT`. The alternative would be to insist
on creating a type alias for anything we wish to use as a Resource.