Simplify nodeUp procedure.
NodeUp procedure could be greatly simplified in case if following
conditions are held:
1. Event producer will try to resend messages until he will be
able to contact EQ.
2. Event queue will try to send messages to the new RC until RC
will acknowledge them.
3. RC will acknowledge NodeUp iff it stored all required information
in ResourceGraph.

All of the conditions above are held in the current halon and are
requirements for the system implementation.

New nodeup procedure does not try to perform and endless loop
of sending messages to the EQ, and just sends promulgate and is
waiting for the result.