d-p-scheduler: Ack SpawnedProcess messages.
forkProcess could return a pid before the scheduler is aware of it.
If the caller blocks immediately, the scheduler could think that the
system is in deadlock.

This patch avoids that by having the scheduler ack the SpawnedProcess
message. Now forkProcess does not return until the caller receives the
ack.