Improve nodeUp procedure
nodeUp first updates the list of EQs of the local EQ tracker,
and then repeatedly tries to send an event to the EQs.

If the EQ tracker is not registered yet, nodeUp will fail to update it.

Current commit changes nodeUp to ensure that EQT is running before
updating it.