Add preferred replicas to the list of replicas.
If a new preffered replica is adverticed to EQT
it will store it in all replicas list, this may
be needed, because  EQT on TS is empty and
preffered replicas didn't update list of all replicas.
As a result because list of all replicas was empty
promulgate didn't send any event.