replicated-log: Protect timeout from async exceptions during clean up.
If timeout was interrupted during clean up, the original exception that
caused to execute the clean up would be ignored. This could be
problematic when calls to timeout are nested and when using 'kill'.

An exception thrown by kill could be overriden by the exception thrown
by an outer call to timeout.