halon: startupHalonNode returns when initialization completes.
Before, tests that used startupHalonNode couldn't know when
initialization was complete, and sometimes ignition could be called too
soon. With this patch, startupHalonNode returns when initialization is
complete.

This in turn required spawning the EQT in a new thread. Therefore, a new
function startEQTracker is used to both spawn the EQTracker and register
it.