Intoduce new primitives for better message handing in RC.
We handle message only after we have done all the work
for that message. In order to do this new method
`startProcessingMsg` was introduced, this method adds
message to a list of the processed messages. Then
another concurrent rule will not use that message.
And finishProcessingMsg that removes the message from
the memory.