replicated-log: do not create storage on start.
When replica is starting it tries to recover it self from the
persistent state even if it doesn't exists. During this process
leveldb creates directory structure for it, this is bad because
we could try to run upper level programs on diskless stations.
Solution is not try to recover the state when directory does not
exist.