RC: improve nodeup rule.
Use new functionality RC in order to not introduce unbounded
memory grows. When new nodeup request arrive we add it to the
set of handled events and process, once we finished we remove it.
This helps in case if concurrent events arrive.
We acknowledge message to EQ only when it's finished, this will
allow EQ to resend message to another instance of RC if current
will die without processing the rule.