halon: Relax na-should-compress-path
This test was failing in CI because messages were asked to arrive
in a particular order to EQs. There was no guarantee provided by the
tested code though. In a slow machine, it turns out messages can arrive
in a different order. Since circleci runs tests rather slowly, we
stop checking the order.