d-p-scheduler: Have runs with the same seed produce the same scheduling.
ProcessIds have a randomly selected and unique integer associated to them.
This unique is not revealed by 'show' but it does affect the order in which
ProcessIds are stored in the internal Maps of the scheduler.

The non-deterministic ordering of ProcessIds causes executions to be
non-deterministic as well. The fix is to set the global random number
generator so the same uniques are produced when given the same seed to
'startScheduler'.