Introduce global worker in rpclite
We need a dedicated thread that we know could perform
mero action, for this reason we introduced a thread
that is started together with mero initialization.
We provide 2 ways: deferrred (postponed) initialization,
when mero is initialized only on first call and in-time
initialization.

Mero-halon make use of this global worker, resulting in
a code simplification.