replicated-log: Preserve requests on lease renewal.
It was possible that a lease renewal was submitted while a proposal was
ongoing. If the lease renewal completes after the next request is
proposed, this could cause the request to try to propose a value for a
decree already taken for the lease renewal.

This was undesirable becase the second request would be discarded and
the client would need to resend it. Also, this was wasteful because the
proposer would try to submit a proposal doomed to fail.

The current patch makes sure lease renewals are submitted only if there
are not ongoing proposals.