Use CApiFFI calling convention to C.
This should make it easier for us to detect breaking changes when things
are modified on the Mero side. This commit itself fixes one issue with
`m0_spiel_tx_commit_forced` using the wrong argument type.

Note that a few places are still using `ccall` - there are places
where Mero requires `const foo` arguments. I know not how to deal
with this in capi.