replicated-log: Write snapshots asynchronously.
This is necessary to have replicas keep the lease and still handle
requests when a snapshot is being written.

When snapshots are small, there is no penalty in writing them
synchronously, but once they are big, the time they take to write
becomes disruptive to the working of the group of replicas.

The snapshot server still handles write requests one at a time, but
now it is careful to check that clients which could be queued for long
are still alive. This allowes to cancel requests to write snapshots by
simply killing the process who submits the write request.