replicated-log: Avoid consecutive requests to write snapshots.
Snapshots are written asynchronously, and requests are canceled as soon as
a new request is sent. The implementation, however, was sending requests too
frequently, and this would cause big messages to accumulate in the mailbox of
the snapshot server.

This patch only has write requests resubmitted when the old request would
persist a snapshot which is too old.