Use Stack's docker support in CI, with a new image.
This commit rolls in two substantial changes: simplify away much of the
custom scripting to build a docker image including dependencies, and
then miraculously getting cabal-install to run inside this docker
container to do all the building and testing, by instead relying on
Stack's first-class support for Docker.

This move made much of the existing docker image content less necessary,
and in any case a side goal was to allow CI to at least check
that *building* with USE_MERO works. This is only possible in a Docker
container that has enough system deps to enable compiling Mero inside
it. Instead of expanding the custom Dockerfile and shell script based
provisioning, the route I took here is to copy and then update the
Puppet scripts for the Dev VM, since the puppet scripts provided much of
that functionality already.

The build workflow is this:

* There is a new repo called seagate-images. Among other things that
repo contains a Dockerfile.
* Whenever the repo gets updated, a new image is automatically generated
and pushed to quay.io, a private docker image registry.
* Then, we setup the halon repo CI so that each build first pulls the
latest image from the registry,
* then builds Mero inside it,
* then builds Halon inside with USE_MERO (including the Haddocks),
* and then again without USE_MERO,
* and then finally runs Halon tests.

The building Mero step requires importing Mero as a submodule, also part
of this PR. We gain precise control regarding what version of mero to
use when building Halon. If you want to use a different version, you
must push a new commit changing the submodule ref.