Run GC of the RG periodically during sync.
Running during sync means that when we're modifying the graph, there
are no surprises about things disappearing at the wrong time. We count
the rough number of ‘disconnect’ calls as a simple heuristic of
whether we want to actually perform the GC: if we're not disconnecting
a lot of stuff, major GC is probably not worth it to begin with.

The few new fields are exposed to the user so they can more finely
control what happens when if they so desire.