Dispatch messages in SSPL service.
Move message dispatching logic from SSPL RC rules to SSPL service.
This makes RC rules simpler as they are interested only in certain
types of the rules. As an additional effect we reduce amount of
pressure on CPU (as we no longer seek in the queue with irrelevant
messages) and memory as we no longer keep messages that are not
needed in the rule queue.