Introduce a way to remove unknown messages from the EQ.
Because RC can't decode message of an unknown type it's not
possible to extract identifier out of such messages. For
this reason we introduce a way to remove messages based on
equality of their serializations. This will be a bit more expensive
then removing messages based on their uuid as it may be needed
to inspect larger amount of data, but such calls expected to be
a rare case.