Don't make shallow clones.
This actually messes up submodule ref checkout, because the refs aren't
necessarily the tip of a branch. This issue wasn't covered up by
previous runs of CircleCI because CircleCI uses a source cache (not to
be confused with the dependency cache), and this source cache has a bug:
it doesn't get invalidated when flags affecting how a history gets
fetched change.