Build and run tests remotely, on an ephemeral AWS machine.
The architecture is this:

1. CircleCI installs Vagrant,
2. provisions a fresh AWS machine using Vagrant,
3. restores the stack build cache on the remote machine,
4. pulls the tweag/halon docker image,
5. builds mero inside the docker image,
6. builds halon and haddocks inside the docker image,
7. runs tests, benchmarks,
8. destroys the AWS machine.

The API tokens for doing this are stored out-of-repo in the CircleCI
web UI configuration. The ssh key though is kept in ci/ci.pem. The AWS
machine is set to self-destruct within 1h, just in case e.g. the build
gets cancelled in the CircleCI web UI.