Update HPI rule.
Now HPI response have information about closure where device is
located, so we could reconstruct RG in case if we do not have initial
data. Despite the fact that initial data load is currently a requirement
this feature is useful, because this will be the only way to add new
enclosures without system re-initialization.

Now HPI request uses existance of WantsReplacement relation, so
now if HPI request contains update that we need to attach device,
DriveInsert event will be emitted, and drive will be attached immediately.