Relocate calls to updateDriveManagerWithFailure
The place it has been called in before would be entered in the case
that we were being notified by the DM in the first place so it was
unnecessary. Instead we call it from various places in the drive
failure rule.

The meaning of the test is changed slightly because we now use custom
rule instead of mimicking an event that would come from SSPL, at least
in part. This more directly tests updateDriveManagerWithFailure but
doesn't reflect a real scenario too well anymore; we have integration
tests elsewhere however and it would greatly complicate matters if we
tried to fully replicate everything as would happen on a cluster so we
settle for this test instead.