replicated-log: Avoid competition of replicas.
Replicas would compete unnecessarily when the notification of a replica
arrives slightly late. A proposal for a decree below the watermark could
be sent which would interfere with the leader.

This patch makes sure proposals are only made for unknown decrees, and
interrupts ongoing proposals if the decrees become known.