Reuse existing nodes on initial data.
This commit fixes initial data load behaviour, so when initial is
loaded it first checks if graph nodes that described in data already
exists in a graph and it that case to not recreate them. This is
needed because graph could be reconstructed based on information from
sspl service, or initial data can be loaded more than once.