mero-halon: use asynchronous synchronization interface.
In mero-halon all RC rules now implement async graph synchronization.
We have few main case:
- synchronization is the last action before rule commit:
  syncGraphProcessMsg is used in that case, rule finishes directly.
  This method could also be used in rules where only Process actions
  are needed.
- rule do not care about synchronization:
  syncGraph (return ()) is used in that case.
- rule that cases about synchronization and need to wait for that to
  happen in this case rule specific data type is introduced and
  syncGraphProcess $ \self -> usend self Specific
  is used.

As currently communication with multimap proces is possible only
from the local process Frontier and SSPL.HL services were fixed
to accomodate that. Now those services have to parts one that
talks to the user and another that works with a graph on RC side.