Fix graph GC
The main problem is that the set of root nodes can not be stored in
the multimap as the multimap is a simple key-value mapping store. We
could either hack the multimap to try to store and then later send a
set of extra metainfo or to store the root nodes themselves in a
format we can already work with: this commit does the latter.

Also see the TODO in ‘sync’ for a potential big problem.