Make graph meta data outside of the graph itself
Keep meta-data inside of the graph is fragile: we expose plenty of
functions with which the user could get hold of it and do things we
don't want the user to do by mistake, such as delete the node holding
the data. Now we put the meta data in replicated state explicitly
instead of piggy-backing on the multimap itself.