Make the GC algorithm ‘breadth-ier’.
As per profiling information (not included), the majority of the time
in the GC algorithm was being spent in taking a difference between the
set of nodes that are candidates for deletions and set of nodes we
know are connected which would then be passed recursively to the next
step of computation. As the deletion candidates start as the set of
all nodes in the graph, this was a costly operation. This difference
was computed for every single node we detected as connected which
added up to a lot of time.

Now instead of performing this difference each time we find a
connected node, we only perform it once per connected node's
connections. In other words, we exhaust the ‘grey’ set in the
algorithm completely instead of processing it one by one. By
processing all the grey nodes at once, we greatly cut down on number
of ‘difference’ calls we have to do. The pay-off is ‘unions’ call but
this is much, much cheaper especially on our data: it's faster to
union ‘n’ sets of connected nodes for ‘n’ grey nodes than to call
‘difference’ at least ‘n * m’ times where ’m’ is the number of nodes
connected to each grey node.

It is easy to guess from the above that we should not be much change
if we have a graph that's deep but not wide. Indeed running the gc-rg
benchmark can confirm this, showing very similar time for the cases
where the nodes are simply connected in sequence. As a result, instead
of having the first 3 major GCs ran during generateConf take ~10-20
seconds on my machine, they now run as follows:

Fri Jan  8 08:16:52 UTC 2016 pid://127.0.0.1:9010:0:50: [RG] Garbage collecting 3529 nodes ...
Fri Jan  8 08:16:52 UTC 2016 pid://127.0.0.1:9010:0:50: [RG] After GC, 2509 nodes remain.
Fri Jan  8 08:16:53 UTC 2016 pid://127.0.0.1:9010:0:50: [RG] Garbage collecting 7841 nodes ...
Fri Jan  8 08:16:53 UTC 2016 pid://127.0.0.1:9010:0:50: [RG] After GC, 4665 nodes remain.
Fri Jan  8 08:16:53 UTC 2016 pid://127.0.0.1:9010:0:50: [RG] Garbage collecting 6325 nodes ...
Fri Jan  8 08:16:53 UTC 2016 pid://127.0.0.1:9010:0:50: [RG] After GC, 5495 nodes remain.
Fri Jan  8 08:16:53 UTC 2016 pid://127.0.0.1:9010:0:50: Loaded initial data