halon: cancel promulgateWait call in case of exception in caller.
Main consumer of the `promulgateWait` call is RecoveryCoordinator,
one part of it's requirements is that all promulgate calls should
be cancelled in case if RC dies because of quorum loss or cancel.
This behaviour also make sense for other uses of the 'promulgateWait',
beacause caller explicitly wait for message to be persisted, so
in case of exception caller can no longer control that.

For old behavoiur one could still use 'promulgate' primitive directly.