Use existing EQ nodes if none specified for halonctl
This is much saner than removing known EQ nodes and then having
promulgate wait for some node in the future to update the list of
nodes so it can send its message. Indeed, in case of `generateConf`,
there is a race condition on when the TS and SAT start: TS was
starting a bit after SAT, setting a meaningful value for EQ nodes
allowing SAT to proceed with `nodeUp`. Now the script should work no
matter of the order of events.