Isolate client and server configuration.
Server conf tree is created during initial data load, and should
be defined only in the YAML file. Client configuration, on the
other hand, is generated by Halon when it detects a new client.

This commit separates client and server configuration and prevents
an issue whereby erroneous configuration could be generated for
a server node.