Rewrite m0d service to handle process control.
This commit rewrites Mero server rules to use process control in m0d.

This was initially intended to address a few issues:
- sysconfig files for 'early' processes were being subsequently overwritten.
- No support for starting multiple 'early' processes.
- Configuration specified by initial data was being ignored in favour of
  defaults.

It's grown into a bit more of a larger refactor through moving all logic
for controlling Mero processes into the m0d service. This removes the
redundancy and (to my mind) makes it easier to comprehend, but
requires some additional work in the rules to handle service starting.

TODO:
- Make sure that appropriate timeouts are included in the server rules.

Fix warnings.

Fix haddock